requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
flask==3.0.2
flask-cors==4.0.0